Purged all components as noted in PR 1374
