+++ src/components/minionList.jsx
@@ -23,52 +23,6 @@ const MinionList = () => {
     }
   }, [daoOverview]);
 
-  // const getMinionData = minion => {
-  //   if (!minion?.minionType) return 'minon';
-  //   switch (minion.minionType) {
-  //     case MINION_TYPES.SUPERFLUID:
-  //       return {
-  //         badgeColor: 'green',
-  //         badgeTextColor: 'white',
-  //         badgeName: 'SF',
-  //         badgeVariant: 'solid',
-  //         url: `/dao/${daochain}/${daoid}/settings/superfluid-minion/${minion.minionAddress}`,
-  //       };
-  //     case MINION_TYPES.UBER:
-  //       return {
-  //         badgeColor: 'purple',
-  //         badgeTextColor: 'white',
-  //         badgeName: 'UHS',
-  //         badgeVariant: 'solid',
-  //         url: `/dao/${daochain}/${daoid}/allies`,
-  //       };
-  //     case MINION_TYPES.NIFTY:
-  //       return {
-  //         badgeColor: 'orange',
-  //         badgeTextColor: 'white',
-  //         badgeName: 'NIFTY',
-  //         badgeVariant: 'solid',
-  //         url: `/dao/${daochain}/${daoid}/vaults/minion/${minion.minionAddress}`,
-  //       };
-  //     case MINION_TYPES.NEAPOLITAN:
-  //       return {
-  //         badgeColor: 'pink',
-  //         badgeTextColor: '#632b16',
-  //         badgeName: 'NEAPOLITAN',
-  //         badgeVariant: 'outline',
-  //         url: `/dao/${daochain}/${daoid}/vaults/minion/${minion.minionAddress}`,
-  //       };
-  //     default:
-  //       return {
-  //         badgeColor: 'white',
-  //         badgeTextColor: 'black',
-  //         badgeName: 'Vanilla',
-  //         badgeVariant: 'solid',
-  //         url: `/dao/${daochain}/${daoid}/vaults/minion/${minion.minionAddress}`,
-  //       };
-  //   }
-  // };
-
   const copiedToast = () => {
     toast({
       title: 'Copied Minion Address',
