--- src/contracts/neapolitanMinionFactory.json
@@ -1,99 +1,0 @@
-[
-  {
-    "inputs": [
-      {
-        "internalType": "address payable",
-        "name": "_template",
-        "type": "address"
-      },
-      {
-        "internalType": "address",
-        "name": "_molochTemplate",
-        "type": "address"
-      }
-    ],
-    "stateMutability": "nonpayable",
-    "type": "constructor"
-  },
-  {
-    "anonymous": false,
-    "inputs": [
-      {
-        "indexed": true,
-        "internalType": "address",
-        "name": "minion",
-        "type": "address"
-      },
-      {
-        "indexed": true,
-        "internalType": "address",
-        "name": "moloch",
-        "type": "address"
-      },
-      {
-        "indexed": false,
-        "internalType": "string",
-        "name": "details",
-        "type": "string"
-      },
-      {
-        "indexed": false,
-        "internalType": "string",
-        "name": "minionType",
-        "type": "string"
-      },
-      {
-        "indexed": false,
-        "internalType": "uint256",
-        "name": "minQuorum",
-        "type": "uint256"
-      }
-    ],
-    "name": "SummonMinion",
-    "type": "event"
-  },
-  {
-    "inputs": [{ "internalType": "uint256", "name": "", "type": "uint256" }],
-    "name": "minionList",
-    "outputs": [{ "internalType": "address", "name": "", "type": "address" }],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "name": "minionType",
-    "outputs": [{ "internalType": "string", "name": "", "type": "string" }],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [{ "internalType": "address", "name": "", "type": "address" }],
-    "name": "minions",
-    "outputs": [
-      { "internalType": "address", "name": "moloch", "type": "address" },
-      { "internalType": "string", "name": "details", "type": "string" }
-    ],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      { "internalType": "address", "name": "moloch", "type": "address" },
-      { "internalType": "string", "name": "details", "type": "string" },
-      { "internalType": "uint256", "name": "minQuorum", "type": "uint256" }
-    ],
-    "name": "summonMinion",
-    "outputs": [{ "internalType": "address", "name": "", "type": "address" }],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "name": "template",
-    "outputs": [
-      { "internalType": "address payable", "name": "", "type": "address" }
-    ],
-    "stateMutability": "view",
-    "type": "function"
-  }
-]
