+++ src/formBuilder/StepperForm.jsx
@@ -9,7 +9,6 @@ import DiscordNotificationsLaunch from './discordLaunchForm';
 import FormBuilder from './formBuilder';
 import Signer from '../components/signer';
 import TheSummoner from '../components/theSummoner';
-import TheLauncher from '../components/theLauncher';
 
 // const getFormCtaText = (currentStep) => {
 //   if()
@@ -179,18 +178,6 @@ const StepperForm = props => {
         boostContent={boostContent}
         secondaryBtn={secondaryBtn}
         handleThen={handleThen}
-      />
-    );
-  }
-  if (currentStep?.type === 'launcher') {
-    return (
-      <TheLauncher
-        {...currentStep}
-        localForm={parentForm}
-        next={currentStep.next}
-        goToNext={goToNext}
-        boostContent={boostContent}
-        secondaryBtn={secondaryBtn}
       />
     );
   }
