+++ src/formBuilder/discordLaunchForm.jsx
@@ -11,8 +11,8 @@ import {
   Text,
 } from '@chakra-ui/react';
 
-import { notificationBoostContent } from '../content/boost-content';
 import { get } from '../utils/requests';
+import { notificationBoostContent } from '../data/boosts';
 
 const DiscordNotificationsLaunch = props => {
   const { stepOverride, goToNext, setStepperStorage, next } = props;
