+++ src/formBuilder/formBuilder.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from 'react';
-import { Flex, FormControl } from '@chakra-ui/react';
 import { useForm } from 'react-hook-form';
 import { Flex, FormControl } from '@chakra-ui/react';
