+++ src/formBuilder/inputFactory.jsx
@@ -24,9 +24,6 @@ import PaymentInput from './paymentInput';
 import PriceInput from './priceInput';
 import RageInput from './rageInput';
 import RaribleNftSelect from './raribleNftData';
-import ColorPicker from './colorPicker';
-import DiscourseMeta from './discourseMeta';
-import ConditionalInput from './conditionalInput';
 import CheckSwitch from './checkSwitch';
 import CheckGate from './checkGate';
 import SuperfluidPaymentInput from './superfluidPaymentInput';
